added error handling to delete method

+++ app.py
@@ -133,21 +133,28 @@ def add_book():
 # Delete operation
 @app.route("/api/v1/book/<int:id>", methods=["DELETE"])
 def delete_book(id):
-    logging.info("Deleting book_id: " + id)
-    try:
-        cursor = db.cursor()
-        query = "DELETE FROM books WHERE book_id = %s"
-        value = (id,)
-        logging.info(query + value)
-        cursor.execute(query, value)
-        db.commit()
-        cursor.close()
-        logging.info("Successfully deleted data from database")
-        return jsonify({"status": "success"}), 204
-    except Exception as e:
-        db.rollback()
-        logging.error("Error occurred while deleting data from database: %s", str(e))
-        return jsonify({"error": str(e)}), 500
+    if id is None:
+        return jsonify({"error": "Book ID is required"}), 400
+    if request.method == "DELETE":
+        logging.info("Deleting book_id: " + id)
+        try:
+            cursor = db.cursor()
+            query = "DELETE FROM books WHERE book_id = %s"
+            value = (id,)
+            logging.info(query + value)
+            cursor.execute(query, value)
+            db.commit()
+            cursor.close()
+            logging.info("Successfully deleted data from database")
+            return jsonify({"status": "success"}), 204
+        except Exception as e:
+            db.rollback()
+            logging.error(
+                "Error occurred while deleting data from database: %s", str(e)
+            )
+            return jsonify({"error": str(e)}), 500
+    else:
+        return jsonify({"error": "Method not allowed"}), 405
 
 
 # # Update operation
